Proyecto2/Redis/requirements.txt
async-timeout==4.0.3
blinker==1.6.3
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
protobuf==4.21.12
redis==5.0.1
simplejson==3.19.2
Werkzeug==3.0.1
